sensemap/sensemap/requirements.txt
opencv-python
torch
torchvision
pcl
pyyaml
rosdep
orb-slam2
yolov5
matplotlib
plotly